requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.6
fonttools==4.56.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
spotipy==2.25.0
sqlparse==0.5.3
threadpoolctl==3.5.0
urllib3==2.3.0
yellowbrick==1.5
